backend/requirements.txt
fastapi
pydantic
uvicorn
strawberry-graphql[fastapi]